Redistributing the ratticweb tests
Was a bit much for one file
