+++ staff/tests.py
@@ -6,7 +6,7 @@ from django.core.urlresolvers import reverse
 from django.test.utils import override_settings
 from django.conf import settings
 
-from ratticweb.tests import TestData
+from ratticweb.tests.helper import TestData
 
 
 class ImportTests(TestCase):
